requirements.txt
gymnasium==0.29.1
numpy
matplotlib
pillow